refactor: Mock Serializer 파일을 더 적절한 위치로 이동
