+++ app/src/app/entry.client.tsx
@@ -3,10 +3,14 @@ import { hydrateRoot } from "react-dom/client";
 import { HydratedRouter } from "react-router/dom";
 import { worker } from "../features/card/mocks/browser";
 
-await worker.start({
-  onUnhandledRequest: "bypass",
-  serviceWorker: { url: "/react-payments/mockServiceWorker.js" },
-});
+const USE_MOCK = import.meta.env.VITE_USE_MOCK === "true";
+
+if (USE_MOCK) {
+  await worker.start({
+    onUnhandledRequest: "bypass",
+    serviceWorker: { url: "/react-payments/mockServiceWorker.js" },
+  });
+}
 
 startTransition(() => {
   hydrateRoot(
